+++ scripts/render_all_themes.sh
@@ -1,0 +1,6 @@
+export SHOW_RICH_ARGPARSE_THEMES=true
+export RENDER_HELP_FORMAT=png
+export RENDER_HELP_OUTPUT_DIR=doc/themes/
+export RENDER_THEME_FILENAME=True
+
+python -m rich_argparse_plus da
